Update how runsc's port-forward command is terminated.
It currently expects all 3 events to happen to exit the command, the container
is killed, local port forward is compeleted, a termination signal is caught.

One can only terminate the command expecting both termination signal
(e.g. ctrl + c) and the container being killed. Instead, the ideal behavior
is that the command exits on either of previous 2 events. The local port
forward goroutine is blocked until the context's cancel function is called.

Thus, we just have the waitgroup to wait for the completion of the local
port forward task.

PiperOrigin-RevId: 663013926